+++ src/app/src/main/kotlin/org/imperial/mrc/hint/logging/GenericLogger.kt
@@ -1,0 +1,6 @@
+package org.imperial.mrc.hint.logging
+
+interface GenericLogger
+{
+    fun info(log: LogMetadata)
+}