+++ src/app/static/src/tests/downloadResults/actions.test.ts
@@ -45,7 +45,7 @@ describe(`download Results actions`, () => {
 
         const state = mockDownloadResultsState();
 
-        mockAxios.onGet(`download/submit/summary/calibrate1`)
+        mockAxios.onPost(`download/submit/summary/calibrate1`)
             .reply(200, mockSuccess(partialDownloadResultsState));
 
         await actions.prepareSummaryReport({commit, state, dispatch, rootState: root} as any);
@@ -55,8 +55,8 @@ describe(`download Results actions`, () => {
         expect(commit.mock.calls[0][0]["payload"]).toEqual(DOWNLOAD_TYPE.SUMMARY)
         expect(commit.mock.calls[1][0]["type"]).toBe("PreparingSummaryReport")
         expect(commit.mock.calls[1][0]["payload"]).toEqual(partialDownloadResultsState)
-        expect(mockAxios.history.get.length).toBe(1);
-        expect(mockAxios.history.get[0]["url"]).toBe("download/submit/summary/calibrate1");
+        expect(mockAxios.history.post.length).toBe(1);
+        expect(mockAxios.history.post[0]["url"]).toBe("download/submit/summary/calibrate1");
 
         expect(dispatch.mock.calls.length).toBe(1)
         expect(dispatch.mock.calls[0]).toEqual(["poll", DOWNLOAD_TYPE.SUMMARY])
@@ -69,7 +69,7 @@ describe(`download Results actions`, () => {
         });
 
         await actions.prepareSummaryReport({commit, state} as any);
-        expect(mockAxios.history.get.length).toBe(0);
+        expect(mockAxios.history.post.length).toBe(0);
         expect(commit.mock.calls.length).toBe(0);
     });
 
@@ -120,7 +120,7 @@ describe(`download Results actions`, () => {
 
         const state = mockDownloadResultsState();
 
-        mockAxios.onGet(`download/submit/summary/calibrate1`)
+        mockAxios.onPost(`download/submit/summary/calibrate1`)
             .reply(500, mockFailure("TEST FAILED"));
 
         await actions.prepareSummaryReport({commit, state, dispatch, rootState: root} as any);
@@ -254,20 +254,23 @@ describe(`download Results actions`, () => {
             modelCalibrate: mockModelCalibrateState({calibrateId: "calibrate1"})
         });
 
+        const getter = {projectState: "json"}
+
         const state = mockDownloadResultsState();
 
-        mockAxios.onGet(`download/submit/spectrum/calibrate1`)
+        mockAxios.onPost(`download/submit/spectrum/calibrate1`, "json")
             .reply(200, mockSuccess(downloadId));
 
-        await actions.prepareSpectrumOutput({commit, state, dispatch, rootState: root} as any);
+        await actions.prepareSpectrumOutput({commit, state, dispatch, rootState: root, rootGetters: getter} as any);
 
         expect(commit.mock.calls.length).toBe(2);
         expect(commit.mock.calls[0][0]["type"]).toBe("SetFetchingDownloadId")
         expect(commit.mock.calls[0][0]["payload"]).toEqual(DOWNLOAD_TYPE.SPECTRUM)
         expect(commit.mock.calls[1][0]["type"]).toBe("PreparingSpectrumOutput")
         expect(commit.mock.calls[1][0]["payload"]).toEqual(downloadId)
-        expect(mockAxios.history.get.length).toBe(1);
-        expect(mockAxios.history.get[0]["url"]).toBe("download/submit/spectrum/calibrate1");
+        expect(mockAxios.history.post.length).toBe(1);
+        expect(mockAxios.history.post[0]["url"]).toBe("download/submit/spectrum/calibrate1");
+        expect(mockAxios.history.post[0]["data"]).toBe("json");
 
         expect(dispatch.mock.calls.length).toBe(1)
         expect(dispatch.mock.calls[0]).toEqual(["poll", DOWNLOAD_TYPE.SPECTRUM])
@@ -280,7 +283,7 @@ describe(`download Results actions`, () => {
         });
 
         await actions.prepareSpectrumOutput({commit, state} as any);
-        expect(mockAxios.history.get.length).toBe(0);
+        expect(mockAxios.history.post.length).toBe(0);
         expect(commit.mock.calls.length).toBe(0);
     });
 
@@ -445,10 +448,12 @@ describe(`download Results actions`, () => {
 
         const state = mockDownloadResultsState();
 
-        mockAxios.onGet(`download/submit/spectrum/calibrate1`)
+        const getter = {projectState: "json"}
+
+        mockAxios.onPost(`download/submit/spectrum/calibrate1`, "json")
             .reply(500, mockFailure("TEST FAILED"));
 
-        await actions.prepareSpectrumOutput({commit, state, dispatch, rootState: root} as any);
+        await actions.prepareSpectrumOutput({commit, state, dispatch, rootState: root, rootGetters: getter} as any);
 
         expect(commit.mock.calls[1][0]).toStrictEqual({
             type: "SpectrumError",
@@ -496,7 +501,7 @@ describe(`download Results actions`, () => {
 
         const state = mockDownloadResultsState();
 
-        mockAxios.onGet(`download/submit/coarse-output/calibrate1`)
+        mockAxios.onPost(`download/submit/coarse-output/calibrate1`)
             .reply(200, mockSuccess(downloadId));
 
         await actions.prepareCoarseOutput({commit, state, dispatch, rootState: root} as any);
@@ -506,8 +511,8 @@ describe(`download Results actions`, () => {
         expect(commit.mock.calls[0][0]["payload"]).toEqual(DOWNLOAD_TYPE.COARSE)
         expect(commit.mock.calls[1][0]["type"]).toBe("PreparingCoarseOutput")
         expect(commit.mock.calls[1][0]["payload"]).toEqual(downloadId)
-        expect(mockAxios.history.get.length).toBe(1);
-        expect(mockAxios.history.get[0]["url"]).toBe("download/submit/coarse-output/calibrate1");
+        expect(mockAxios.history.post.length).toBe(1);
+        expect(mockAxios.history.post[0]["url"]).toBe("download/submit/coarse-output/calibrate1");
 
         expect(dispatch.mock.calls.length).toBe(1)
         expect(dispatch.mock.calls[0]).toEqual(["poll", DOWNLOAD_TYPE.COARSE])
@@ -520,7 +525,7 @@ describe(`download Results actions`, () => {
         });
 
         await actions.prepareCoarseOutput({commit, state} as any);
-        expect(mockAxios.history.get.length).toBe(0);
+        expect(mockAxios.history.post.length).toBe(0);
         expect(commit.mock.calls.length).toBe(0);
     });
 
@@ -680,7 +685,7 @@ describe(`download Results actions`, () => {
 
         const state = mockDownloadResultsState();
 
-        mockAxios.onGet(`download/submit/coarse-output/calibrate1`)
+        mockAxios.onPost(`download/submit/coarse-output/calibrate1`)
             .reply(500, mockFailure("TEST FAILED"));
 
         await actions.prepareCoarseOutput({commit, state, dispatch, rootState: root} as any);
@@ -732,7 +737,7 @@ describe(`download Results actions`, () => {
 
         const state = mockDownloadResultsState();
 
-        mockAxios.onGet(`download/submit/comparison/calibrate1`)
+        mockAxios.onPost(`download/submit/comparison/calibrate1`)
             .reply(200, mockSuccess(downloadId));
 
         await actions.prepareComparisonOutput({commit, state, dispatch, rootState: root} as any);
@@ -742,8 +747,8 @@ describe(`download Results actions`, () => {
         expect(commit.mock.calls[0][0]["payload"]).toEqual(DOWNLOAD_TYPE.COMPARISON)
         expect(commit.mock.calls[1][0]["type"]).toBe("PreparingComparisonOutput")
         expect(commit.mock.calls[1][0]["payload"]).toEqual(downloadId)
-        expect(mockAxios.history.get.length).toBe(1);
-        expect(mockAxios.history.get[0]["url"]).toBe("download/submit/comparison/calibrate1");
+        expect(mockAxios.history.post.length).toBe(1);
+        expect(mockAxios.history.post[0]["url"]).toBe("download/submit/comparison/calibrate1");
 
         expect(dispatch.mock.calls.length).toBe(1)
         expect(dispatch.mock.calls[0]).toEqual(["poll", DOWNLOAD_TYPE.COMPARISON])
@@ -756,7 +761,7 @@ describe(`download Results actions`, () => {
         });
 
         await actions.prepareComparisonOutput({commit, state} as any);
-        expect(mockAxios.history.get.length).toBe(0);
+        expect(mockAxios.history.post.length).toBe(0);
         expect(commit.mock.calls.length).toBe(0);
     });
 
@@ -923,7 +928,7 @@ describe(`download Results actions`, () => {
 
         const state = mockDownloadResultsState();
 
-        mockAxios.onGet(`download/submit/comparison/calibrate1`)
+        mockAxios.onPost(`download/submit/comparison/calibrate1`)
             .reply(500, mockFailure("TEST FAILED"));
 
         await actions.prepareComparisonOutput({commit, state, dispatch, rootState: root} as any);